+++ stroom-proxy/stroom-proxy-app/src/main/java/stroom/proxy/app/App.java
@@ -72,9 +72,11 @@ public class App extends Application<Config> {
     @Override
     public void initialize(final Bootstrap<Config> bootstrap) {
         // This allows us to use templating in the YAML configuration.
-        bootstrap.setConfigurationSourceProvider(new SubstitutingSourceProvider(
-                bootstrap.getConfigurationSourceProvider(),
-                new EnvironmentVariableSubstitutor(false)));
+        bootstrap.setConfigurationSourceProvider(
+                new ProxyConfigurationSourceProvider(
+                        new SubstitutingSourceProvider(
+                                bootstrap.getConfigurationSourceProvider(),
+                                new EnvironmentVariableSubstitutor(false))));
 //        bootstrap.addBundle(new AssetsBundle("/ui", ResourcePaths.ROOT_PATH, "index.html", "ui"));
     }