+++ stroom-proxy/stroom-proxy-app/src/main/java/stroom/proxy/app/ContentSyncConfig.java
@@ -1,12 +1,16 @@
 package stroom.proxy.app;
 
+import stroom.util.shared.IsProxyConfig;
+
 import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 
 import java.util.Map;
 import javax.inject.Singleton;
 
 @Singleton
-public class ContentSyncConfig {
+@JsonPropertyOrder(alphabetic = true)
+public class ContentSyncConfig implements IsProxyConfig {
 
     private boolean isContentSyncEnabled = false;
     private Map<String, String> upstreamUrl;