+++ stroom-proxy/stroom-proxy-app/src/main/java/stroom/proxy/app/RestClientConfig.java
@@ -1,0 +1,16 @@
+package stroom.proxy.app;
+
+import stroom.util.shared.IsProxyConfig;
+
+import com.fasterxml.jackson.annotation.JsonPropertyOrder;
+import io.dropwizard.client.JerseyClientConfiguration;
+
+import javax.inject.Singleton;
+
+@Singleton
+@JsonPropertyOrder(alphabetic = true)
+public class RestClientConfig extends JerseyClientConfiguration implements IsProxyConfig {
+
+    // This class makes it easier for us to use/bind the JerseyClientConfiguration in our config tree
+    // and makes it easier to find its usages in IJ
+}