+++ stroom-proxy/stroom-proxy-app/src/main/java/stroom/proxy/app/handler/ProxyRequestConfig.java
@@ -1,8 +1,15 @@
 package stroom.proxy.app.handler;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
+import stroom.util.shared.IsProxyConfig;
 
-public class ProxyRequestConfig {
+import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.annotation.JsonPropertyOrder;
+
+import javax.inject.Singleton;
+
+@Singleton
+@JsonPropertyOrder({"storingEnabled", "repoDir", "format", "rollCron"})
+public class ProxyRequestConfig implements IsProxyConfig {
 
     private String receiptPolicyUuid;